pyproject.toml
[project]
name = "aspdc-certificates"
version = "0.1.0"
description = "Certificate generator for ASPDC events with QR code support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pypdf2>=3.0.0",
    "reportlab>=4.0.0",
    "qrcode[pil]>=7.4.2",
    "pillow>=10.0.0",
]

[project.urls]
Repository = "https://github.com/aspdc/certificate-generator"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
